// LDC.Atlas.WebApp/Client/app/shared/enums/costmatrix-parameters-fields.enum.ts
export enum CostMatrixParametersFields {
    Cmy1 = 'Cmy1',
    Cmy2 = 'Cmy2',
    Cmy3 = 'Cmy3',
    Cmy4 = 'Cmy4',
    Cmy5 = 'Cmy5',
    ContractTerms = 'Contract terms',
    ContractType = 'Contract type',
    Counterparty = 'Counterparty',
    PaymentTerms = 'Payment terms',
    PortOfDestination = 'Port of destination',
    PortOfOrigin = 'Port of origin',
}